commiting 0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,4 +1,5 @@
 #include "lists.h"
+#include<stdio.h>
 /**
  * print_list-prints contents of a linked list
  * @h:passed list head
@@ -7,12 +8,12 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t count = 0;
+	unsigned int count = 0;
 
 	while (h != NULL)
 	{
 		if (h->str)
-			printf("[%d] %s\n", h->len, h->str);
+			printf("[%u] %s\n", h->len, h->str);
 		else
 			printf("[0] (nil)\n");
 		h = h->next;
